{{cookiecutter.repository}}/requirements/production.txt
# Pro-tip: Try not to put anything here. There should be no dependency in
#   production that isn't in development.
-r common.txt

# Static Files and Media Storage
# -------------------------------------
django-storages==1.7.1
boto3==1.9.165

# Caching
# -------------------------------------
django-redis==4.10.0
hiredis==1.0.0

{% if cookiecutter.newrelic == 'y' %}
# Logging
# -------------------------------------
newrelic==4.20.0.120
{% endif %}

{%- if cookiecutter.add_django_auth_wall.lower() == 'y' %}
# Basic Auth Protection
# -------------------------------------
django-auth-wall==0.2.0
{%- endif %}
